+++ models/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [6.0.0]
+### Removed
+ - encodeAccessToken and decodeAccessToken functions
+### Added
+ - accessTokenEncoder and loginTokenEncoder Encoder objects
 
 ## [5.1.0]
 ### Added